Start single frame when plunger released
as opposed to when it is first pressed. It is with great consternation
that I make this change. The original intention of the Trigger()
feature was to be instant, but having used the intervalometer on
several rolls now, I find the extra frame that this feature introduces
into sequences to be troublesome.